Adding more user friendly exceptions when building the mirroring inventory

+++ lib/minimart/mirror/inventory.rb
@@ -1,15 +1,17 @@
 require 'yaml'
-require 'fileutils'
 
 module Minimart
   class Mirror
+    class InvalidInventoryError < Exception; end
+
     class Inventory
 
       attr_reader :inventory_config,
-                  :inventory_directory,
-                  :downloaded_cookbooks
+                  :inventory_directory
 
       def initialize(opts = {})
+        raise ArgumentError.new('missing required :inventory_config option') if opts[:inventory_config].nil?
+
         @inventory_config    = parse_config_file(opts[:inventory_config])
         @inventory_directory = opts[:inventory_directory]
       end
@@ -22,18 +24,21 @@ module Minimart
       private
 
       def make_inventory_directory
-        FileUtils.mkdir_p inventory_directory
+        Utils::FileHelper.make_directory(inventory_directory)
       end
 
       def download_cookbooks!
-        Minimart::Configuration.output.puts "Building inventory..."
+        Configuration.output.puts "Building inventory..."
 
-        @downloaded_cookbooks ||= base_source.download_cookbooks do |cookbook, archived_cookbook|
-          destination = "#{inventory_directory}/#{cookbook.name}-#{cookbook.version}"
-          Minimart::Utils::Archive.extract_cookbook(archived_cookbook, destination)
+        base_source.download_cookbooks do |cookbook, archived_cookbook|
+          Utils::Archive.extract_cookbook(archived_cookbook, inventory_destination(cookbook))
         end
 
-        Minimart::Configuration.output.puts "Done building inventory..."
+        Configuration.output.puts "Done building inventory..."
+      end
+
+      def inventory_destination(cookbook)
+        "#{inventory_directory}/#{cookbook.name}-#{cookbook.version}"
       end
 
       def base_source
@@ -41,6 +46,8 @@ module Minimart
       end
 
       def parse_config_file(path)
+        raise InvalidInventoryError.new('The inventory configuration file could not be found') unless Utils::FileHelper.file_exists?(path)
+
         YAML.load(File.open(path))
       end
